some basic tests for services save methods

+++ src/main/java/com/example/footballmadrid/Repositories/CommentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommentRepository extends JpaRepository<CommentModel, Integer> {
+public interface CommentRepository extends JpaRepository<CommentModel, Long> {
 }
