+++ src/main/java/com/example/footballmadrid/Repositories/GameRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GameRepository extends JpaRepository<GameModel, Integer> {
+public interface GameRepository extends JpaRepository<GameModel, Long> {
 }
